ロガーの共通処理を抽象クラスに凝集
