+++ samples/web-csr/dressca-backend/system-common/build.gradle
@@ -20,6 +20,7 @@ repositories {
 dependencies {
 	implementation supportDependencies.spring_boot_starter
 	implementation supportDependencies.commons_lang3
+	implementation supportDependencies.spring_boot_starter_log4j2
 	testImplementation supportDependencies.spring_boot_starter_test
 }
