+++ samples/web-csr/dressca-backend/system-common/src/main/java/com/dressca/systemcommon/log/DresscaLogger.java
@@ -1,18 +1,33 @@
 package com.dressca.systemcommon.log;
 
 import java.util.Map;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import com.dressca.systemcommon.constant.SystemPropertyConstants;
 
 /**
  * コンテキストにログを保持させる機能を追加したカスタムロガーのインターフェースです。
  */
-public interface DresscaLogger {
+public abstract class DresscaLogger {
+
+  private static final Logger apLog = LoggerFactory.getLogger(SystemPropertyConstants.APPLICATION_LOG_LOGGER);
+
+  /**
+   * 構造化ログでデフォルトで出力する内容を追加します。
+   * try ブロック内に MDC.put(key, value) で追記してください。
+   * 
+   * @param logAction ログを出力するメソッド。
+   */
+  protected abstract void logWithMdc(Runnable logAction);
 
   /**
    * DEBUG レベルのログを出力します。
    * 
    * @param msg ログのメッセージ。
    */
-  public void debug(String msg);
+  public void debug(String msg) {
+    logWithMdc(() -> apLog.debug(msg));
+  }
 
   /**
    * Key Value を追加して DEBUG レベルのログを出力します。
@@ -20,14 +35,22 @@ public interface DresscaLogger {
    * @param msg         ログのメッセージ。
    * @param keyValueMap Key Value のマップ。
    */
-  public void debug(String msg, Map<String, String> keyValueMap);
+  public void debug(String msg, Map<String, String> keyValueMap) {
+    logWithMdc(() -> {
+      var logBuilder = apLog.atDebug().setMessage(msg);
+      keyValueMap.forEach(logBuilder::addKeyValue);
+      logBuilder.log();
+    });
+  }
 
   /**
    * ERROR レベルのログを出力します。
    * 
    * @param msg ログのメッセージ。
    */
-  public void error(String msg);
+  public void error(String msg) {
+    logWithMdc(() -> apLog.error(msg));
+  }
 
   /**
    * Key Value を追加して ERROR レベルのログを出力します。
@@ -35,14 +58,23 @@ public interface DresscaLogger {
    * @param msg         ログのメッセージ。
    * @param keyValueMap Key Value のマップ。
    */
-  public void error(String msg, Map<String, String> keyValueMap);
+  public void error(String msg, Map<String, String> keyValueMap) {
+
+    logWithMdc(() -> {
+      var logBuilder = apLog.atError().setMessage(msg);
+      keyValueMap.forEach(logBuilder::addKeyValue);
+      logBuilder.log();
+    });
+  }
 
   /**
    * INFO レベルのログを出力します。
    * 
    * @param msg ログのメッセージ。
    */
-  public void info(String msg);
+  public void info(String msg) {
+    logWithMdc(() -> apLog.info(msg));
+  }
 
   /**
    * Key Value を追加して INFO レベルのログを出力します。
@@ -50,14 +82,22 @@ public interface DresscaLogger {
    * @param msg         ログのメッセージ。
    * @param keyValueMap Key Value のマップ。
    */
-  public void info(String msg, Map<String, String> keyValueMap);
+  public void info(String msg, Map<String, String> keyValueMap) {
+    logWithMdc(() -> {
+      var logBuilder = apLog.atInfo().setMessage(msg);
+      keyValueMap.forEach(logBuilder::addKeyValue);
+      logBuilder.log();
+    });
+  }
 
   /**
    * TRACE レベルのログを出力します。
    * 
    * @param msg ログのメッセージ。
    */
-  public void trace(String msg);
+  public void trace(String msg) {
+    logWithMdc(() -> apLog.trace(msg));
+  }
 
   /**
    * Key Value を追加して TRACE レベルのログを出力します。
@@ -65,14 +105,23 @@ public interface DresscaLogger {
    * @param msg         ログのメッセージ。
    * @param keyValueMap Key Value のマップ。
    */
-  public void trace(String msg, Map<String, String> keyValueMap);
+  public void trace(String msg, Map<String, String> keyValueMap) {
+    logWithMdc(() -> {
+      var logBuilder = apLog.atTrace().setMessage(msg);
+      keyValueMap.forEach(logBuilder::addKeyValue);
+      logBuilder.log();
+    });
+  }
 
   /**
    * WARN レベルのログを出力します。
    * 
    * @param msg ログのメッセージ。
    */
-  public void warn(String msg);
+  public void warn(String msg) {
+
+    logWithMdc(() -> apLog.warn(msg));
+  }
 
   /**
    * Key Value を追加して WARN レベルのログを出力します。
@@ -80,5 +129,12 @@ public interface DresscaLogger {
    * @param msg         ログのメッセージ。
    * @param keyValueMap Key Value のマップ。
    */
-  public void warn(String msg, Map<String, String> keyValueMap);
+  public void warn(String msg, Map<String, String> keyValueMap) {
+    logWithMdc(() -> {
+      var logBuilder = apLog.atWarn().setMessage(msg);
+      keyValueMap.forEach(logBuilder::addKeyValue);
+      logBuilder.log();
+    });
+  }
+
 }
